+++ src/oss2db/kill2db.cpp
@@ -43,14 +43,14 @@ void Kill2Db::process(const OaStatStruct &oss)
 	string target = ""; //Parameter 1
 	string killer = ""; //Parameter 0
 	//mod is parameter 2
-	if(oss.parameters.at(1)>-1 && oss.parameters.at(1)<clientIdMap.size() && oss.parameters.at(1)!=1022 )
+	if(oss.parameters.at(1)>-1 && oss.parameters.at(1)<(int)clientIdMap.size() && oss.parameters.at(1)!=1022 )
 	{
 		target = clientIdMap.at(oss.parameters.at(1));
 	}
 	else
 		target ="WORLD"; //Should not be possible... except maybe for NPC-creatures
 
-	if(oss.parameters.at(0)>-1 && oss.parameters.at(0)<clientIdMap.size() && oss.parameters.at(0)!=1022)
+	if(oss.parameters.at(0)>-1 && oss.parameters.at(0)<(int)clientIdMap.size() && oss.parameters.at(0)!=1022)
 	{
 		killer = clientIdMap.at(oss.parameters.at(0));
 	}
